refacto: updated connection validation
